@@ -5,7 +5,8 @@
 	import { currentPostData } from '$lib/utils/stores.js';
 	import { page } from '$app/stores';
 	import { processContent } from '$lib/utils';
-	export let data;
+	/** @type {{data: any}} */
+	let { data } = $props();
 	currentPostData.set({ category: data.meta.category, path: $page.url.pathname });
 	/**@type {(s:string|number|Date)=>(string)}*/
 	let toISO = (s) => {
@@ -136,7 +137,7 @@
 		</div>
 	{/if}
 	<div class="content" use:processContent>
-		<svelte:component this={data.content} />
+		<data.content />
 		<a href={data.meta.link} target="_blank" class="cta"
 			>{data.meta.link_text ?? 'Ir a su página'}</a
 		>
