@@ -1,12 +1,12 @@
 <script>
-	import {page} from "$app/stores"
+	import {page} from "$app/state"
 </script>
 <main>
 	<a href="/">⬅️ volver a la página</a>
 
 	<form method="POST">
 		<input type="submit" value="Iniciar sesión con GitHub"/>
-		<input hidden type="text" name="redirectTo" value={$page.url.searchParams.get('redirectTo') ?? '/'} />
+		<input hidden type="text" name="redirectTo" value={page.url.searchParams.get('redirectTo') ?? '/'} />
 	</form>
 </main>
 <style>
